Starting a new game
To start a new game after someone has already been playing we need to:

1. Remove all current tiles on the board
2. Place two new starter tiles on the board
3. Start up all of the systems used to play the game

Removing all tiles on the board is something Bevy makes fairly direct. We can create a new `game_reset` system that queries for all entities with a `Position` component. If there are any, we need to despawn them, so we loop over the iterator to operate on each tile.

`commands.entity` gives us an `EntityCommands` builder that we can use to operate on the given entity. In this case Bevy has a `despawn` and a `despawn_recursive` method. `despawn` will remove the current entity from the game, while `despawn_recursive` will remove the current entity _and_ all of it's children. In this case we want `despawn_recursive` because we want to remove the `Text` inside the tile as well.

```rust
fn game_reset(
    mut commands: Commands,
    tiles: Query<Entity, With<Position>>,
    mut game: ResMut<Game>,
) {
    for entity in tiles.iter() {
        commands.entity(entity).despawn_recursive();
    }
    game.score = 0;
}
```

We'll additionally reset the game score to 0.

in `src/main.rs` in the app builder we'll add this system as a new set tied to the `on_enter` hook for the `RunState::Playing` state. This means that when we enter the `Playing` state, the `game_reset` system will run once, just like a startup system.

```rust
.add_system_set(
    SystemSet::on_enter(RunState::Playing)
        .with_system(
            game_reset.system(),
        )
)
```

If you run the game now, you'll notice one big issue. The `game_reset` system is running _after_ the `spawn_tiles` system, resulting in no tiles on the board when we start a new game!

We can remove the `spawn_tiles` system in the `PostStartup` hook.

```rust
.add_startup_system_to_stage(
    StartupStage::PostStartup,
    spawn_tiles.system(),
)
```

because we're going to add it after the `game_reset` system now.

```rust
.add_system_set(
    SystemSet::on_enter(RunState::Playing)
        .with_system(game_reset.system())
        .with_system(spawn_tiles.system()),
)
```

This works!

Because of the way commands are applied (deferred to the end of each stage) and what we're changing in these systems, we can run both of these systems in parallel and the `game_reset` system will only remove the pre-existing tiles while the `spawn_tiles` system will only add new tiles to the board.

Running the game now will allow you to reset and restart the game.